Config exception handling & crash fixes
Fixed a crash when the global config was invalid
Fixed a crash when trying to continue using an already loaded file after selecting an incompatible file or a file that causes an exception when reading config
Config will now provide clear reasons for issues
Program will show an issue with global config and then close without crashing
M4A config will now properly free resources when closing the ROM

+++ VG Music Studio/Core/GBA/M4A/M4APlayer.cs
@@ -28,7 +28,7 @@ namespace Kermalis.VGMusicStudio.Core.GBA.M4A
             thread.Start();
         }
 
-        public void LoadSong(int index)
+        public void LoadSong(long index)
         {
             SongEntry entry = config.Reader.ReadObject<SongEntry>(config.SongTableOffsets[0] + (index * 8));
             SongHeader header = config.Reader.ReadObject<SongHeader>(entry.HeaderOffset - GBAUtils.CartridgeOffset);
